Player gets animator and weaponholder from their skin now
The player in the Game Scene has no model currently. If you launch
the game through the Title now, you get assigned a skin, and scripts use
the skin's animator. I added a WeaponHolder transform to all the prefabs in
the TT prefabs folder and am essentially using those as the skin models.
Backed up the previous player model as TT_demo_zombie_player if we still need it.